requirements.txt
google-genai
numpy
Pillow
tqdm
